+++ src/main/scala/org/podval/tools/test/Test.scala
@@ -1,6 +1,7 @@
 package org.podval.tools.test
 
-import org.gradle.api.internal.tasks.testing.TestDescriptorInternal
+import org.gradle.api.internal.tasks.testing.{TestCompleteEvent, TestDescriptorInternal, TestResultProcessor, TestStartEvent}
+import org.gradle.api.tasks.testing.TestResult.ResultType
 
 abstract class Test extends TestDescriptorInternal:
   final override def getParent: Test = null
@@ -8,3 +9,27 @@ abstract class Test extends TestDescriptorInternal:
   final override def getClassDisplayName: String = getClassName
 
   def getParentId: Object
+
+  def started(
+    time: Long,
+    testResultProcessor: TestResultProcessor
+  ): Unit = testResultProcessor.started(
+    this,
+    TestStartEvent(
+      time,
+      getParentId
+    )
+  )
+  
+  def completed(
+    time: Long,
+    resultType: ResultType,
+    testResultProcessor: TestResultProcessor
+  ): Unit = testResultProcessor.completed(
+    getId,
+    TestCompleteEvent(
+      time,
+      resultType
+    )
+  )
+