+++ src/main/scala/org/podval/tools/test/TestSerializerRegistry.scala
@@ -1,13 +1,16 @@
 package org.podval.tools.test
 
 import org.gradle.api.internal.tasks.testing.{TestCompleteEvent, TestDescriptorInternal, TestStartEvent}
-import org.gradle.api.tasks.testing.TestOutputEvent
+import org.gradle.api.tasks.testing.{TestOutputEvent, TestFailure}
 import org.gradle.api.tasks.testing.TestResult.ResultType
 import org.gradle.internal.serialize.{BaseSerializerFactory, DefaultSerializerRegistry, Serializer, SerializerRegistry}
 import org.podval.tools.test.serializer.*
 import sbt.testing.Selector
 
 // Note: based on org.gradle.api.internal.tasks.testing.worker.TestEventSerializer
+// TODO maybe I can *add* my classes to the registries instead of copying and modifying Gradle classes?
+// And if not, I should file a pull request against Gradle to make registries overridable...
+// TODO can I throw an exception when the class is not serializable?
 final class TestSerializerRegistry extends DefaultSerializerRegistry:
   private val baseSerializerFactory: BaseSerializerFactory = new BaseSerializerFactory
 
@@ -50,13 +53,14 @@ final class TestSerializerRegistry extends DefaultSerializerRegistry:
   register(classOf[TestStartEvent], TestStartEventSerializer(nullableIdSerializer))
   register(classOf[TestCompleteEvent], TestCompleteEventSerializer(nullableResultTypeSerializer))
   register(classOf[TestOutputEvent], TestOutputEventSerializer(outputDestinationSerializer))
+  register(classOf[TestFailure], TestFailureSerializer(throwableSerializer))
 
   register(classOf[Throwable], throwableSerializer)
 
-// Note: org.gradle.internal.remote.internal.hub.DefaultMethodArgsSerializer
+  // Note: org.gradle.internal.remote.internal.hub.DefaultMethodArgsSerializer
   // seems to make a decision based on the type of the first argument of a method
   // (and somehow I didn't even see calls to canSerialize() for those...);
-  // in TestResultProcessor, first argument is often an id - so, I attracting attention:
+  // in TestResultProcessor, first argument is often an id - so, I am attracting attention:
   register(classOf[Object], idSerializer)
 
   // Note: some TestResultProcessor subclasses wrap things in TestDescriptorInternal,