+++ src/main/scala/org/podval/tools/test/WorkerTestClassProcessor.scala
@@ -1,7 +1,8 @@
 package org.podval.tools.test
 
-import org.gradle.api.internal.tasks.testing.{JULRedirector, TestClassProcessor, TestClassRunInfo, TestResultProcessor}
-import org.gradle.api.internal.tasks.testing.processors.CaptureTestOutputTestResultProcessor
+import org.gradle.api.internal.tasks.testing.{TestClassProcessor, TestClassRunInfo, TestResultProcessor}
+import org.gradle.api.internal.tasks.testing.processors.{CaptureTestOutputTestResultProcessor,
+  DefaultStandardOutputRedirector, StandardOutputRedirector}
 
 // Note: translated from org.gradle.api.internal.tasks.testing.worker.WorkerTestClassProcessor and modified:
 // - not to wrap everything in WorkerTestSuiteDescriptor;
@@ -13,7 +14,7 @@ import org.gradle.api.internal.tasks.testing.processors.CaptureTestOutputTestRes
 //   - use CaptureTestOutputTestResultProcessor to capture output.
 //
 // Now that forking of the tests works, this can be rolled back to the original,
-// but it does not seem worth it:
+// but it does not seem to be worth it:
 // - I'll need serializers for WorkerTestSuiteDescriptor and CompositeId (where both scope and id are Longs);
 // - I do not think I want the per-worker synthetic suites;
 // - I already have parentId for my suites, which I'll need to null out to conform to the expectations of the Gradle code.
@@ -21,11 +22,11 @@ final class WorkerTestClassProcessor(processor: TestClassProcessor) extends Test
 
   override def startProcessing(testResultProcessor: TestResultProcessor): Unit =
     val resultProcessor: TestResultProcessor =
-//        CaptureTestOutputTestResultProcessor(
-          testResultProcessor
-//          ,
-//          new JULRedirector
-//        )
+//      CaptureTestOutputTestResultProcessor(
+        testResultProcessor
+//        ,
+//        new DefaultStandardOutputRedirector // TODO JUL?
+//      )
     processor.startProcessing(resultProcessor)
 
   override def processTestClass(testClass: TestClassRunInfo): Unit =