+++ test-projects/frameworks/src/test/scala/org/podval/tools/test/MUnitTest.scala
@@ -1,8 +1,12 @@
 package org.podval.tools.test
 
 class MUnitTest extends munit.FunSuite:
-  test("hello") {
+  test("42 != 43") {
     val obtained = 42
     val expected = 43
     assertEquals(obtained, expected)
+  }
+
+  test("2=2") {
+    assertEquals(2, 2)
   }