+++ test-projects/frameworks/src/test/scala/org/podval/tools/test/ScalaCheckTest.scala
@@ -3,7 +3,7 @@ package org.podval.tools.test
 import org.scalacheck.Properties
 import org.scalacheck.Prop.forAll
 
-object ScalaCheckTest /*StringSpecification*/ extends Properties("String") {
+class ScalaCheckTest /*StringSpecification*/ extends Properties("String") {
 
   property("startsWith") = forAll { (a: String, b: String) =>
     (a+b).startsWith(a)