+++ test-projects/frameworks/src/test/scala/org/podval/tools/test/Spec2Test.scala
@@ -11,9 +11,11 @@ final class Spec2Test /*HelloWorldSpec*/ extends Specification:
     contain 11 characters $e1
     start with 'Hello' $e2
     end with 'world' $e3
+    fail to end with 'xxx' $e4
 
     """
 
     def e1 = "Hello world" must haveSize(11)
     def e2 = "Hello world" must startWith("Hello")
     def e3 = "Hello world" must endWith("world")
+    def e4 = "Hello world" must endWith("xxx")